[GGB-127] MemberCreate,Reader 분리
